initial 정립중
