+++ lib/screens/initial/viewmodel/initial_page_viewmodel_abs.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+// Initial Page ViewModel
+abstract class InitialPageViewModel implements Listenable {
+  void goLoginPage(BuildContext context);
+  void goRegisterPage(BuildContext context);
+}
