+++ lib/screens/login/login_screen.dart
@@ -1,52 +1,138 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:go_router/go_router.dart';
 
-import 'package:dahae_mobile/common/util/route_animation.dart';
 import 'package:dahae_mobile/screens/login/login_component.dart';
-import 'package:dahae_mobile/screens/login/sign_in_screen.dart';
-import 'package:dahae_mobile/screens/register/sign_up_email.dart';
 
-class LogInScreen extends StatelessWidget {
-  const LogInScreen({super.key});
+class LoginScreen extends StatefulWidget {
+  @override
+  State<LoginScreen> createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  bool _isWrong = false;
+  final formKey = GlobalKey<FormState>();
+  //final emailController = TextEditingController();
+  String email = '';
+  String password = '';
+  FocusNode _emailFocus = FocusNode();
+  FocusNode _passwordFocus = FocusNode();
 
   @override
   Widget build(BuildContext context) {
-    return WillPopScope(
-      onWillPop: () async => false,
-      child: Scaffold(
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
+    //WARNING MESSAGE
+    Center wrongAccountMsg = const Center(
+      child: Text('잘 못 입력하셨습니다.',
+          style: TextStyle(
+              color: Color(0xFFFA0000),
+              fontSize: 12,
+              fontWeight: FontWeight.w500)),
+    );
+
+    SingleChildScrollView screenBody = SingleChildScrollView(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          SizedBox(
+              height: MediaQuery.of(context).viewInsets.bottom == 0
+                  ? MediaQuery.of(context).size.height * 0.17
+                  : MediaQuery.of(context).size.height * 0.1),
+          const Image(
+              image: AssetImage('assets/images/logo_3d.png'), height: 130),
+          const SizedBox(height: 25),
+          LoginPageText(),
+          SizedBox(height: 30, child: _isWrong ? wrongAccountMsg : Container()),
+          SignUpInputTextBox(
+            label: '이메일',
+            focusNode: _emailFocus,
+            onSaved: (val) {
+              setState(() {
+                email = val;
+              });
+            },
+            onChanged: (val) {},
+            validator: (val) => CheckValidate().validateEmail(_emailFocus, val),
+          ),
+          const SizedBox(height: 10),
+          SignUpInputTextBox(
+            label: '비밀번호',
+            password: true,
+            nopad: true,
+            focusNode: _passwordFocus,
+            onSaved: (val) {
+              setState(() {
+                password = val;
+              });
+            },
+            onChanged: (val) {},
+            validator: (val) => CheckValidate().validateEmpty(_emailFocus, val),
+          ),
+          const SizedBox(height: 10),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.end,
             children: [
-              const Image(
-                  image: AssetImage('assets/images/logo_3d.png'), height: 150),
-              const SizedBox(height: 25),
-              const LoginPageText(),
-              const SizedBox(height: 70),
-              DahaeStartButton(
-                child: Text('처음 다 해 봐요!'),
-                onPressed: () {
-                  PageRouteWithAnimation pageRouteWithAnimation =
-                      PageRouteWithAnimation(SignUpScreen());
-                  Navigator.push(
-                      context, pageRouteWithAnimation.slideRitghtToLeft());
-                },
+              GestureDetector(
+                onTap: () => GoRouter.of(context).go('/habit'),
+                child: Text(
+                  '비밀번호가 생각나지 않아요',
+                  style: TextStyle(fontSize: 9),
+                ),
               ),
-              const SizedBox(height: 20),
-              DahaeStartButton(
-                child: Text('다 해 본 적 있어요'),
-                onPressed: () {
-                  PageRouteWithAnimation pageRouteWithAnimation =
-                      PageRouteWithAnimation(SignInScreen());
-                  Navigator.push(
-                      context, pageRouteWithAnimation.slideRitghtToLeft());
-                  //MaterialPageRoute(
-                  //builder: (_) => SignInScreen(),),
-                },
-              ),
+              SizedBox(height: 30, width: 30),
             ],
           ),
-        ),
+          const SizedBox(height: 50)
+        ],
       ),
     );
+
+    SignUpBottomButton bottomButton = SignUpBottomButton(
+      text: '로그인',
+      onPressed: () async {
+        if (formKey.currentState?.validate() == true) {
+          formKey.currentState?.save();
+          // DB로 보내서 맞는지 확인
+          if (true) {
+            GoRouter.of(context).go('/habit');
+          } else {
+            _isWrong = true;
+          }
+        }
+      },
+    );
+
+    return Form(
+      key: formKey,
+      child: Scaffold(
+          //resizeToAvoidBottomInset: false,
+          backgroundColor: Colors.white,
+          appBar: const PreferredSize(
+              preferredSize: Size.fromHeight(kToolbarHeight),
+              child: SignUpAppBar()),
+          extendBodyBehindAppBar: true,
+          body: screenBody,
+          // sign in button
+          bottomSheet: bottomButton),
+    );
+  }
+}
+
+class SignUpScreen2 extends StatefulWidget {
+  const SignUpScreen2({
+    super.key,
+    required this.email,
+  });
+
+  final String email;
+
+  @override
+  State<SignUpScreen2> createState() => _SignUpScreen2State();
+}
+
+class _SignUpScreen2State extends State<SignUpScreen2> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
   }
 }
