+++ lib/screens/screens.dart
@@ -2,5 +2,8 @@ export 'home_scaffold.dart';
 export 'habit/habit_screen.dart';
 export 'pet/pet_screen.dart';
 export 'profile/profile_screen.dart';
+export 'initial/page/initial_page.dart';
+export 'splash/splash_screen.dart';
+
+export 'register/register_email.dart';
 export 'login/login_screen.dart';
-export 'login/splash_screen.dart';
